# api/openapi/api-gateway.yaml
openapi: 3.1.0
info:
  title: Personal Finance OS API Gateway
  version: 0.4.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/profile:
    get:
      summary: Get current authenticated profile from JWT claims
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated profile
  /auth/login:
    post:
      summary: Exchange username and password for a JWT pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
  /auth/refresh:
    post:
      summary: Refresh access token using a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refreshed JWT pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
  /auth/me:
    get:
      summary: Inspect current subject through auth-service
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current authenticated subject
  /imports/raw:
    post:
      summary: Upload a raw statement import into MongoDB and enqueue parse work
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Raw import accepted
  /imports/{import_id}:
    get:
      summary: Get raw import status from MongoDB
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: import_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Raw import metadata
  /parser/results/{import_id}:
    get:
      summary: Get parsed import projection from MongoDB
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: import_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Parsed import projection
  /api/v1/transactions:
    get:
      summary: List transactions from the ledger service
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Transactions list
    post:
      summary: Store a transaction in the ledger service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        '201':
          description: Transaction created
  /api/v1/categories:
    get:
      summary: List ledger categories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Categories list
  /api/v1/recurring:
    get:
      summary: Detect recurring transactions from the current ledger slice
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Recurring patterns
  /api/v1/notifications/status:
    get:
      summary: Inspect notification worker status and Telegram configuration
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification worker status
  /api/v1/notifications/preferences:
    get:
      summary: Get current user's notification preferences
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification preferences
    put:
      summary: Update current user's notification preferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                telegram_enabled:
                  type: boolean
                batch_non_critical:
                  type: boolean
                quiet_hours_enabled:
                  type: boolean
                quiet_start_minute:
                  type: integer
                quiet_end_minute:
                  type: integer
                quiet_timezone:
                  type: string
                disabled_alert_types:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Notification preferences updated
  /api/v1/notifications/telegram/demo:
    post:
      summary: Queue a demo Telegram notification job
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Demo notification queued
  /api/v1/notifications/telegram/poll/once:
    post:
      summary: Trigger one Telegram polling cycle
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Telegram polling cycle completed
  /api/v1/notifications/telegram/link/confirm:
    post:
      summary: Confirm Telegram chat binding with a one-time code under JWT
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Telegram chat linked
  /api/v1/analytics/projections:
    get:
      summary: Inspect analytics projection metadata
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Analytics projection metadata
  /api/v1/analytics/projections/daily-spend:
    get:
      summary: Query daily spend projections
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
        - in: query
          name: category
          schema:
            type: string
      responses:
        '200':
          description: Daily spend projection
  /api/v1/analytics/projections/alerts:
    get:
      summary: Query alert projections
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
        - in: query
          name: severity
          schema:
            type: string
      responses:
        '200':
          description: Alert projection
  /api/v1/analytics/projections/summary:
    get:
      summary: Query summary analytics projection
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Summary projection
  /api/v1/presence:
    get:
      summary: Inspect realtime presence state
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Presence snapshot
  /api/v1/realtime/config:
    get:
      summary: Inspect realtime gateway configuration
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Realtime gateway configuration
  /ws:
    get:
      summary: WebSocket entrypoint proxied to realtime-gateway
      description: Use an access token in the Authorization header or `access_token` query parameter.
      security:
        - bearerAuth: []
      responses:
        '101':
          description: WebSocket upgrade successful
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
    Transaction:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        account_id:
          type: string
        merchant:
          type: string
        category:
          type: string
        amount_cents:
          type: integer
        currency:
          type: string
        occurred_at:
          type: string
          format: date-time